I fixed compilation under Windows with MSVC by adding the missing support for dllexport/dllimport.

+++ libs/nyast/BaseClassLibrary/SmallInteger.cpp
@@ -71,7 +71,7 @@ int64_t SmallInteger::asInt64() const
 
 double SmallInteger::asFloat64() const
 {
-    return self().decodeSmallInteger();
+    return double(self().decodeSmallInteger());
 }
 
 Oop SmallInteger::additionWith(Oop other)
